requirements.txt
appdirs==1.4.4
attrs==20.2.0
bcrypt==3.2.0
black==20.8b1
capstone==4.0.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
idna==2.10
iniconfig==1.0.1
intervaltree==3.1.0
Mako==1.1.3
MarkupSafe==1.1.1
mypy==0.782
mypy-extensions==0.4.3
packaging==20.4
paramiko==2.7.2
pathspec==0.8.0
pluggy==0.13.1
psutil==5.7.2
pwntools==4.2.1
py==1.9.0
pycparser==2.20
pyelftools==0.26
Pygments==2.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.4
PySocks==1.7.1
pytest==6.1.0
python-dateutil==2.8.1
regex==2020.9.27
requests==2.24.0
ROPGadget==6.3
six==1.15.0
sortedcontainers==2.2.2
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
unicorn==1.0.2rc3
urllib3==1.25.10
